getDeviceInfo接口修改

+++ src/main/java/com/hust/ewsystem/DAO/PO/Tasks.java
@@ -14,7 +14,7 @@ import java.time.LocalDateTime;
 public class Tasks {
 
     @TableId(type = IdType.AUTO)
-    private Integer taskId;
+    private Long taskId;
 
     private String taskLabel;
